==> notebooks/01_2_zinc_prep.ipynb <==
from rdkit import Chem
import numpy as np
import pandas as pd
from joblib import Parallel, delayed

def make_smile_canonical(smile): # To avoid duplicates, for example: canonical '*C=C(*)C' == '*C(=C*)C'
    try:
        mol = Chem.MolFromSmiles(smile)
        canon_smile = Chem.MolToSmiles(mol, canonical=True)
        # if smile != canon_smile:
        #     print(f'{smile} > {canon_smile}')
        return canon_smile
    except:
        return np.nan

def process_chunk(df_chunk: pd.DataFrame) -> pd.DataFrame:
    # 복사해서 변경(원본 공유로 인한 사이드이펙트 방지)
    out = df_chunk.copy()
    out["canonical_smiles"] = out["smiles"].map(make_smile_canonical)
    return out

ZINC_df = pd.read_parquet('../data/raw/ZINC_cleaned.parquet', columns=['smiles'])
ZINC_df

n_jobs = 16
chunks = np.array_split(ZINC_df, n_jobs)

processed_chunks = Parallel(
    n_jobs=n_jobs,
    backend="loky",
    prefer="processes",
    batch_size=10_000,
)(
    delayed(process_chunk)(chk) for chk in chunks
)

ZINC_df = pd.concat(processed_chunks, ignore_index=True)

ZINC_df.to_parquet(
    '../data/pretrain/ZINC_canonical_cleaned.parquet',
    engine='pyarrow',
    compression='snappy',
    index=False
) # <-- 수정된 부분

ZINC_df.head()
